Move patching code to LinuxImageHeapProvider from CEntryPointSnippets.
